+++ spec/create_gem/detection/runtime_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-RSpec.describe CreateGem::RuntimeVersions::Detector do
+RSpec.describe CreateGem::Detection::Runtime do
   it 'returns ruby, rubygems and bundler versions' do
-    bundler_detector = instance_double(CreateGem::BundlerVersion::Detector, detect!: Gem::Version.new('4.0.4'))
+    bundler_detector = instance_double(CreateGem::Detection::BundlerVersion, detect!: Gem::Version.new('4.0.4'))
     versions = described_class.new(bundler_detector: bundler_detector).detect!
 
     expect(versions.ruby).to be_a(Gem::Version)